Renamed custom module to match the repository For now on, any given custom modules need to be flagged as a generikit-* module.
